now can compile with -Weverything

+++ src/convert/convert_from_short.c
@@ -3,7 +3,7 @@
 
 void convert_from_short(t_fmt *fmt, t_list *buffer, va_list *args) {
   char str[7];
-  const short number = va_arg(*args, int);
+  const short number = (short)va_arg(*args, int);
 
   if (fmt->flag & PLUS && number > -1) {
     ft_itoa(number, str + 1, DECIMAL);
